slf4j collect logs for tr and move settings to init file

+++ src/test/java/core/helpers/framework/general/FileHelper.java
@@ -5,13 +5,25 @@ import java.io.File;
 public class FileHelper {
 
     public static void deleteDirectory(File directory) {
+        if (directory == null || !directory.exists() || !directory.isDirectory()) {
+            System.out.println("Provided directory is null, does not exist, or is not a directory.");
+            return;
+        }
+
         for (File file : directory.listFiles()) {
             if (file.isDirectory()) {
                 deleteDirectory(file);
             } else {
-                file.delete();
+                boolean deleted = file.delete();
+                if (!deleted) {
+                    System.out.println("Failed to delete file: " + file.getAbsolutePath());
+                }
             }
         }
-        directory.delete();
+
+        boolean deleted = directory.delete();
+        if (!deleted) {
+            System.out.println("Failed to delete directory: " + directory.getAbsolutePath());
+        }
     }
 }
